modify to return IActionResult instead of objects, better error handling on API

+++ AWSServerlessAPI/Controllers/ItemsController.cs
@@ -17,8 +17,8 @@ namespace AWSServerlessAPI.Controllers
         private readonly IMapper _mapper;
         private readonly IConfigurationRepository _config;
 
-        public ItemsController(IItemRepository repository, 
-            IMapper mapper, 
+        public ItemsController(IItemRepository repository,
+            IMapper mapper,
             IConfigurationRepository config)
         {
             _repository = repository;
@@ -29,7 +29,7 @@ namespace AWSServerlessAPI.Controllers
         // GET api/values
         [HttpGet]
         public async Task<IEnumerable<ItemModel>> Get()
-        {            
+        {
             var results = await _repository.GetRecords();
 
             var mapped = _mapper.Map<IEnumerable<ItemModel>>(results);
@@ -39,13 +39,18 @@ namespace AWSServerlessAPI.Controllers
 
         // GET api/values/5
         [HttpGet("{id}")]
-        public async Task<ItemModel> Get(string id)
+        public async Task<IActionResult> Get(string id)
         {
             var result = await _repository.GetRecordByKey(id);
 
+            if (result == null)
+            {
+                return NotFound();
+            }
+
             var mapped = _mapper.Map<ItemModel>(result);
 
-            return mapped;
+            return Ok(mapped);
         }
 
         // GET api/values/5
@@ -57,38 +62,64 @@ namespace AWSServerlessAPI.Controllers
 
         // POST api/values
         [HttpPost]
-        public async Task<ItemModel> Post([FromBody]ItemModel newRecord)
+        public async Task<IActionResult> Post([FromBody]ItemModel newRecord)
         {
+            if (String.IsNullOrEmpty(newRecord.Id))
+            {
+                return BadRequest("ID is required");
+            }
+
+            var existing = await _repository.GetRecordByKey(newRecord.Id);
+
+            if(existing != null)
+            {
+                return BadRequest("ID already exists");
+            }
+
             var mappedRecord = Mapper.Map<ItemRecord>(newRecord);
 
             var result = await _repository.InsertRecord(mappedRecord);
 
-            return Mapper.Map<ItemModel>(result);
+            return Ok(Mapper.Map<ItemModel>(result));
         }
 
         // PUT api/values/5
         [HttpPut("{id}")]
-        public async Task<ItemModel> Put(string id, [FromBody]ItemModel newRecord)
+        public async Task<IActionResult> Put(string id, [FromBody]ItemModel newRecord)
         {
-            if(id != newRecord.Id)
+            if (id != newRecord.Id)
             {
-                throw new Exception("Route ID does not match payload ID");
+                return BadRequest("Route ID does not match payload ID");
+            }
+
+            var existing = await _repository.GetRecordByKey(newRecord.Id);
+
+            if (existing == null)
+            {
+                return NotFound();
             }
 
             var mappedRecord = Mapper.Map<ItemRecord>(newRecord);
 
             var result = await _repository.UpdateExistingRecord(mappedRecord);
 
-            return Mapper.Map<ItemModel>(result);
+            return Ok(Mapper.Map<ItemModel>(result));
         }
 
         // DELETE api/values/5
         [HttpDelete("{id}")]
-        public async Task Delete(string id)
-        {            
+        public async Task<ActionResult> Delete(string id)
+        {
+            var existing = await _repository.GetRecordByKey(id);
+
+            if (existing == null)
+            {
+                return NotFound();
+            }
+
             await _repository.DeleteRecordByKey(id);
 
-            return;
+            return Ok();
         }
     }
 }
